front-end/src/app/shared/desplegable-tipo-grano/tipo-grano.service.spec.ts
import { TestBed } from '@angular/core/testing';

import { TipoGranoService } from './tipo-grano.service';
import { TestModule } from '../../core/mocks/test.module';
import { ApiService } from '../../core/services/restClient/api.service';
import { configureTestSuite } from '../../core/mocks/testing';
import { TipoGrano } from '../data-models/tipo-grano';
import { of } from 'rxjs/internal/observable/of';

describe('TipoGranoService', () => {
  let apiService: any;
  let service: TipoGranoService;

  configureTestSuite(() => {
    apiService = jasmine.createSpyObj('ApiService', ['get']);
    TestBed.configureTestingModule({
      imports: [TestModule],
      providers: [ TipoGranoService,
      {provide: ApiService, useValue: apiService}]
    });
    service = TestBed.get(TipoGranoService);
  });

  beforeEach(() => {
    apiService.get.calls.reset();
  });

  it('should be created', () => {
    expect(service).toBeTruthy();
  });

  describe('El metodo getAll()', () => {
    it('invoca al metodo get del ApiService', () => {
      service.getAll();

      expect(apiService.get).toHaveBeenCalledWith('tipo-grano');
      expect(apiService.get).toHaveBeenCalledTimes(1);
    });

    it('devuelve un observable con los datos obtenidos por el apiservice.get', () => {
      const esperado = of([{} as TipoGrano]);
      apiService.get.and.returnValue(esperado);

      const resultado = service.getAll();

      expect(resultado).toEqual(esperado);
    });
  });
});
